Add generic client support
